+++ packages/twig-to-php-parser/__tests__/fixtures/larva.config.js
@@ -2,5 +2,6 @@ module.exports = {
 	parser: {
 		twigDir: `${__dirname}/src/patterns/`,
 		phpDir: `${__dirname}/template-parts/patterns/`,
+		isUsingPlugin: true,
 	}
 };