+++ packages/twig-to-php-parser/lib/twig-to-php-parser.php
@@ -63,13 +63,13 @@ function twig_to_php_parser( $patterns_dir_path, $template_dir_path, $is_using_p
 		// https://regex101.com/r/lceCnz/2/
 		$loop_regex = '/({% for item in\s*)(\w*)/';
 
-		// Get matches for {{ name }}
-		// https://regex101.com/r/ACN0rE/2
-		$mustache_regex = '/({{\s*)(\w*?\|?\w*)(\s*}})/';
+		// Get matches for {{ name }}, {{ name|filter }}, {{ name['item'] }}
+		// https://regex101.com/r/ACN0rE/5
+		$mustache_regex = '/({{\s*)(\w*?\[?\'?\w*\'?\]?\|?\w*)(\s*}})/';
 
-		// Get matches for {% include "path/c-element.twig" with data %}
-		// https://regex101.com/r/ns5kBR/2
-		$include_regex = '/({%\sinclude ")(.*[c-|o-|l-].*\/)(.*)(.twig)(" with )(\w*)(.*\s%})/';
+		// Get matches for {% include "path/c-element.twig" with data['some_data'] %}
+		// https://regex101.com/r/ns5kBR/3
+		$include_regex = '/({%\sinclude ")(.*[c-|o-|l-].*\/)(.*)(.twig)(" with )(\w*?\[?\'?\w*\'?\]?\|?\w*)(\s%})/';
 
 		// Get matches for {% include '/path/here/' ~ variable_svg ~ '.svg' %}
 		// https://regex101.com/r/pdleQb/3